requirements.txt
pandas==2.1.4
openpyxl==3.1.2
streamlit==1.29.0
google-genai>=0.8.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
chromadb==0.4.18
numpy==1.24.3